cleanup and enhanced symlinking

+++ build.sh
@@ -13,7 +13,6 @@
 # @todo	    Major: Can only compile libgcrypt wth "-arch i386 -arch x86_64" and not "-arch i386"
 # @todo     Major: We still need sudo commands (key word: $prefix_install.bak)
 # @todo     Major: Enhance autofix and environment testing
-# @todo     Minor: on error do 'sudo mv "$prefix_install.bak" "$prefix_install"'
 # @todo     Minor: clang can not compile gnupg
 # @todo		Minor: Compile libgcrypt without --disable-aesni-support
 # @todo		Minor: Based on installation activate ppc or not
@@ -27,7 +26,7 @@
 #     2. Download Xcode 4: http://itunes.apple.com/us/app/xcode/id448457090?mt=12
 #     3. Install only Xcode 3 Essentials > Xcode Toolset to /Xcode3 (export COMMAND_LINE_INSTALL=1; open "/Volumes/Xcode and iOS SDK/Xcode and iOS SDK.mpkg")
 #     4. Install Xcode 4
-#     5. Run build.sh autofix or:
+#     5. Run "./build.sh autofix" or:
 #        cd /Developer/SDKs/MacOSX10.5.sdk/usr/lib/gcc
 #        sudo ln -s i686-apple-darwin11 i686-apple-darwin10
 #        sudo ln -s powerpc-apple-darwin11 powerpc-apple-darwin10
@@ -79,6 +78,85 @@ export LDFLAGS="-L$prefix_install/lib"
 export LD_LIBRARY_PATH="$prefix_install/lib"
 export LIBUSB_1_0_CFLAGS="-I$prefix_install/include/libusb-1.0/"
 export LIBUSB_1_0_LIBS="-L$prefix_install/lib"
+################################################################################
+
+
+# setup ########################################################################
+mkdir -p "$prefix_build/lib"
+cp "$rootPath/Keys.gpg" "$buildDir/pubring.gpg"
+
+## clean
+if [ "$1" == "clean" ]; then
+    echo -n " * Cleaning..."
+    rm -f "$ccache"
+    rm -rf "$prefix_build"
+    rm -f "$LOGPATH/"*.log
+    rm -rf "$iconv_build/$iconv_version"
+    rm -rf "$gettext_build/$gettext_version"
+    rm -rf "$pth_build/$pth_version"
+    rm -rf "$libusb_build/$libusb_version"
+    rm -rf "$libusbcompat_build/$libusbcompat_version"
+    rm -rf "$libgpgerror_build/$libgpgerror_version"
+    rm -rf "$libassuan_build/$libassuan_version"
+    rm -rf "$libgcrypt_build/$libgcrypt_version"
+    rm -rf "$libksba_build/$libksba_version"
+    rm -rf "$zlib_build/$zlib_version"
+    rm -rf "$gpg_build/$gpg_version"
+	echo " OK"
+	exit 0
+fi
+################################################################################
+
+## autofix
+if [ "$1" == "autofix" ]; then
+    echo " * Autofixing..."
+    echo -n "   * 10.5 SDK: "
+    [ ! -e "$xcode4sdk105" ] && sudo ln -s "$xcode3sdk105" "$xcode4sdk105"
+    echo "OK"
+
+    echo -n "   * Assembler: "
+    [ ! -L "$xcode4as" ] && [ ! -e "$xcode4as.bak" ] && sudo mv "$xcode4as" "$xcode4as.bak"
+    [ ! -L "$xcode4as" ] && [ -e "$xcode4as.bak" ] && sudo ln -Fs "$xcode3as" "$xcode4as"
+    echo "OK"
+
+    echo -n "   * GCC (i686): "
+    [ ! -L "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin11" ] && sudo ln -s "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin10" "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin11"
+    echo "OK"
+
+    echo -n "   * GCC (ppc): "
+    [ ! -L "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin11" ] && sudo ln -s "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin10" "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin11"
+    echo "OK"
+
+	exit 0
+fi
+################################################################################
+
+## testing environment
+echo " * Logfiles: $LOGPATH/build-xyz.log"
+echo " * Target: $prefix_build"
+echo " * Testing environment..."
+
+echo -n "   * 10.5 SDK: "
+[ -e "$xcode4sdk105" ] && echo "OK"
+[ ! -e "$xcode4sdk105" ] && [ -e "$xcode3sdk105" ] && echo "run '$0 autofix'" # && exit 1
+[ ! -e "$xcode4sdk105" ] && [ ! -e "$xcode3sdk105" ] && echo "FAILED. Install Xcode3 first" # && exit 1
+
+echo -n "   * Assembler: "
+[ -L "$xcode4as" ] && echo "OK"
+[ ! -L "$xcode4as" ] && [ -e "$xcode3as" ] && echo "run '$0 autofix'" # && exit 1
+[ ! -L "$xcode4as" ] && [ ! -e "$xcode3as" ] && echo "FAILED. Install Xcode3 first" # && exit 1
+
+echo -n "   * GCC (i386): "
+[ -L "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin11" ] && echo "OK"
+[ ! -L "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin11" ] && echo "run '$0 autofix'" # && exit 1
+
+echo -n "   * GCC (ppc): "
+[ -L "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin11" ] && echo "OK"
+[ ! -L "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin11" ] && echo "run '$0 autofix'" # && exit 1
+
+#echo -n "   * GCC (compile test): "
+#echo "main() {return 0;}" | gcc $CFLAGS -xc -o /dev/null - 2>$LOGFILE
+#if [ "$?" == 0 ]; then echo "OK"; else echo "FAIL (see $LOGFILE)"; exit 1; fi
 ################################################################################
 
 
@@ -174,95 +252,16 @@ gpg_patch="gnupg/AllInOne.patch"
 ################################################################################
 
 
-## init
-if [ "$1" == "clean" ]; then
-    echo -n " * Cleaning..."
-    rm -f "$ccache"
-    rm -rf "$prefix_build"
-    rm -f "$LOGPATH/"*.log
-    rm -rf "$iconv_build/$iconv_version"
-    rm -rf "$gettext_build/$gettext_version"
-    rm -rf "$pth_build/$pth_version"
-    rm -rf "$libusb_build/$libusb_version"
-    rm -rf "$libusbcompat_build/$libusbcompat_version"
-    rm -rf "$libgpgerror_build/$libgpgerror_version"
-    rm -rf "$libassuan_build/$libassuan_version"
-    rm -rf "$libgcrypt_build/$libgcrypt_version"
-    rm -rf "$libksba_build/$libksba_version"
-    rm -rf "$zlib_build/$zlib_version"
-    rm -rf "$gpg_build/$gpg_version"
-	echo " OK"
-	exit 0
-fi
-################################################################################
-
-## autofix
-if [ "$1" == "autofix" ]; then
-    echo " * Autofixing..."
-    echo -n "   * 10.5 SDK: "
-    [ ! -e "$xcode4sdk105" ] && sudo ln -s "$xcode3sdk105" "$xcode4sdk105"
-    echo "OK"
-
-    echo -n "   * Assembler: "
-    [ ! -L "$xcode4as" ] && [ ! -e "$xcode4as.bak" ] && sudo mv "$xcode4as" "$xcode4as.bak"
-    [ ! -L "$xcode4as" ] && [ -e "$xcode4as.bak" ] && sudo ln -s "$xcode3as" "$xcode4as"
-    echo "OK"
-
-    echo -n "   * GCC (i686): "
-    [ ! -L "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin11" ] && sudo ln -s "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin10" "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin11"
-    echo "OK"
-
-    echo -n "   * GCC (ppc): "
-    [ ! -L "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin11" ] && sudo ln -s "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin10" "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin11"
-    echo "OK"
-
-	exit 0
-fi
-
-## testing environment
-echo " * Logfiles: $LOGPATH/build-xyz.log"
-echo " * Target: $prefix_build"
-echo " * Testing environment..."
-
-echo -n "   * 10.5 SDK: "
-[ -e "$xcode4sdk105" ] && echo "OK"
-[ ! -e "$xcode4sdk105" ] && [ -e "$xcode3sdk105" ] && echo "run '$0 autofix'" # && exit 1
-[ ! -e "$xcode4sdk105" ] && [ ! -e "$xcode3sdk105" ] && echo "FAILED. Install Xcode3 first" # && exit 1
-
-echo -n "   * Assembler: "
-[ -L "$xcode4as" ] && echo "OK"
-[ ! -L "$xcode4as" ] && [ -e "$xcode3as" ] && echo "run '$0 autofix'" # && exit 1
-[ ! -L "$xcode4as" ] && [ ! -e "$xcode3as" ] && echo "FAILED. Install Xcode3 first" # && exit 1
-
-echo -n "   * GCC (i386): "
-[ -L "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin11" ] && echo "OK"
-[ ! -L "$xcode4sdk105/usr/lib/gcc/i686-apple-darwin11" ] && echo "run '$0 autofix'" # && exit 1
-
-echo -n "   * GCC (ppc): "
-[ -L "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin11" ] && echo "OK"
-[ ! -L "$xcode4sdk105/usr/lib/gcc/powerpc-apple-darwin11" ] && echo "run '$0 autofix'" # && exit 1
-
-#echo -n "   * GCC (compile test): "
-#echo "main() {return 0;}" | gcc $CFLAGS -xc -o /dev/null - 2>$LOGFILE
-#if [ "$?" == 0 ]; then echo "OK"; else echo "FAIL (see $LOGFILE)"; exit 1; fi
-
-################################################################################
-
-
-## ugly hack ;)
-if [ -L "$prefix_install" ]; then
-	sudo rm "$prefix_install"
-elif [ -e "$prefix_install" ]; then
-	sudo mv "$prefix_install" "$prefix_install.bak"
-fi
-
-mkdir -p "$prefix_build/lib"
-sudo ln -Fs "$prefix_build" "$prefix_install"
-cp "$rootPath/Keys.gpg" "$buildDir/pubring.gpg"
-################################################################################
-
-
 # functions ####################################################################
+
+# ugly hack - to be removed
+function setUnsetSymLinks {
+  echo " * Changing MacGPG2 links (might need sudo password)..."
+  [ -L "$prefix_install" ] && sudo rm "$prefix_install"
+  [ ! -d "$prefix_install.bak" ] && [ -d "$prefix_install" ] && sudo mv "$prefix_install" "$prefix_install.bak"
+  [ ! -d "$prefix_install.bak" ] && [ ! -d "$prefix_install" ] && sudo ln -Fs "$prefix_build" "$prefix_install"
+  [ -d "$prefix_install.bak" ] && sudo mv "$prefix_install.bak" "$prefix_install"
+}
 
 function setLogPipe {
     LOGFILE="$LOGPATH/build${1:+-$1}.log"
@@ -275,6 +274,7 @@ function errExit {
 	resetLogPipe
 	echo "$1"
 	echo "See $LOGFILE for details."
+	setUnsetSymLinks
 	exit 1
 }
 function waitfor {
@@ -367,6 +367,12 @@ function install {
 
     resetLogPipe
 }
+################################################################################
+
+
+# init #########################################################################
+# to be removed
+setUnsetSymLinks
 ################################################################################
 
 
@@ -512,10 +518,5 @@ else
 	touch ".checked"
 	resetLogPipe
 fi
-
-if sudo rm -f "$prefix_install"; then
-	if [ -e "$prefix_install.bak" ]; then
-		sudo mv "$prefix_install.bak" "$prefix_install"
-	fi
-fi
+setUnsetSymLinks
 ################################################################################
